ajout de parenté dans gameobject, réalisation des room à portes paramétrées
